Testing to Staging

See merge request 123milhas/finances/api!471

+++ app/Http/Controllers/ChartOfAccountsController.php
@@ -7,6 +7,7 @@ use App\Exports\ChartOfAccountsExport;
 use Illuminate\Http\Request;
 use App\Http\Requests\StoreChartOfAccountsRequest;
 use App\Http\Requests\PutChartOfAccountsRequest;
+use App\Models\ChartOfAccounts;
 use App\Services\ChartOfAccountsService as ChartOfAccountsService;
 
 class ChartOfAccountsController extends Controller
@@ -42,6 +43,12 @@ class ChartOfAccountsController extends Controller
 
     public function destroy($id)
     {
+        if(ChartOfAccounts::where('parent', $id)->exists())
+        {
+            return response()->json([
+                'erro' => 'Este plano de contas está associado a outros planos de contas é necessário apagar e/ou alterar suas dependências antes de apagá lo.'
+            ], 422);
+        }
         $chartOfAccounts = $this->chartOfAccountsService->deleteChartOfAccounts($id);
         return response('');
     }
